Several changes:
Added a module to manage save files.

Added options to save and continue a game, as well as a module for saving and loading a

Typed method parameters with the exception of the self parameter common to all methods.

Added comments in methods and various parts of the code

Removed a deprecated Levels method